automations/ai-content-summariser/requirements.txt
openai>=1.30.0
trafilatura>=1.9.0
flask>=3.0.0
python-dotenv>=1.0.0
